requirements.txt
Django>=4.2.0,<5.0
psycopg2-binary>=2.9.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-extensions>=3.2.0
requests>=2.28.0
python-decouple>=3.6
django-cors-headers>=4.0.0
gunicorn>=20.1.0